Split executor logic to implemnentation
